src/main/java/lotto/domain/lotto/Lotto.java
package lotto.domain.lotto;

import lotto.global.validator.LottoValidator;

import java.util.Collections;
import java.util.List;

public class Lotto {
    private final List<Integer> numbers;

    private Lotto(List<Integer> numbers) {
        LottoValidator.validate(numbers);
        this.numbers = numbers;
        Collections.sort(this.numbers);
    }

    public static Lotto from(final List<Integer> numbers) {
        return new Lotto(numbers);
    }

    public List<Integer> getNumbers() {
        return Collections.unmodifiableList(numbers);
    }
}
